circle: be explicit about various potential data loss conversions

+++ libnopegl/src/node_circle.c
@@ -75,14 +75,14 @@ static int circle_init(struct ngl_node *node)
         goto end;
     }
 
-    const float step = TAU_F32 / o->npoints;
+    const float step = TAU_F32 / (float)o->npoints;
 
     vertices[0] = 0.0f;
     vertices[1] = 0.0f;
     uvcoords[0] = 0.5f;
     uvcoords[1] = 0.5f;
     for (int i = 1; i < nb_vertices; i++) {
-        const float angle = (i - 1) * -step;
+        const float angle = (float)(i - 1) * -step;
         const float x = sinf(angle) * o->radius;
         const float y = cosf(angle) * o->radius;
         vertices[i*3 + 0] = x;
@@ -90,8 +90,8 @@ static int circle_init(struct ngl_node *node)
         uvcoords[i*2 + 0] = (x + 1.0f) / 2.0f;
         uvcoords[i*2 + 1] = (1.0f - y) / 2.0f;
         indices[(i - 1) * 3 + 0]  = 0; // point to center coordinate
-        indices[(i - 1) * 3 + 1]  = i;
-        indices[(i - 1) * 3 + 2]  = i + 1;
+        indices[(i - 1) * 3 + 1]  = (uint16_t)i;
+        indices[(i - 1) * 3 + 2]  = (uint16_t)i + 1;
     }
     /* Fix overflowing vertex reference back to the start for sealing the
      * circle */
